+++ src/scripts/settings.js
@@ -12,6 +12,8 @@ const store = new Store({
     playBackControl: true,
     muteArtists: false,
     mutedArtists: ["TIDAL"],
+    skipArtists: false,
+    skippedArtists: [""],
     disableBackgroundThrottle: true,
     menuBar: true,
     apiSettings: {